dcp-815 separate gcs.storage & gcs.transfer classes
